archive/stage1-legacy/requirements.txt
# Prompt-Scribe Stage 1: Local Data Pipeline
# Python 3.11+ required

# Core Dependencies
pydantic>=2.0.0              # Data validation and settings management
pydantic-settings>=2.0.0     # Settings management with Pydantic

# Database
sqlite-utils>=3.35.0         # SQLite utilities

# Data Processing
pandas>=2.0.0                # Data manipulation
numpy>=1.24.0                # Numerical computing

# LLM Integration (Optional - Data Layer)
openai>=1.0.0                # OpenAI API client
anthropic>=0.7.0             # Anthropic API client

# Utilities
python-dotenv>=1.0.0         # Environment variable management
tqdm>=4.65.0                 # Progress bars
loguru>=0.7.0                # Enhanced logging

# Development Dependencies
pytest>=7.4.0                # Testing framework
pytest-cov>=4.1.0            # Coverage plugin for pytest
pytest-asyncio>=0.21.0       # Async support for pytest
black>=23.7.0                # Code formatter
pylint>=2.17.0               # Code linter
mypy>=1.5.0                  # Static type checker
isort>=5.12.0                # Import sorting

# Documentation
mkdocs>=1.5.0                # Documentation generator
mkdocs-material>=9.2.0       # Material theme for MkDocs

